Shortcuts were enabled even if their corresponding buttons were not

+++ frontend-html/src/model/actions-ui/DataView/TableView/onFieldKeyDown.ts
@@ -41,6 +41,10 @@ import { onCopyRowClick } from "model/actions-ui/DataView/onCopyRowClick";
 import { onFilterButtonClick } from "model/actions-ui/DataView/onFilterButtonClick";
 import { onEscapePressed } from "model/actions-ui/DataView/onEscapePressed";
 import { getFormScreenLifecycle } from "model/selectors/FormScreen/getFormScreenLifecycle";
+import { geScreenActionButtonsState } from "model/actions-ui/ScreenToolbar/saveButtonVisible";
+import { getIsAddButtonVisible } from "model/selectors/DataView/getIsAddButtonVisible";
+import { getIsDelButtonVisible } from "model/selectors/DataView/getIsDelButtonVisible";
+import { getIsCopyButtonVisible } from "model/selectors/DataView/getIsCopyButtonVisible";
 
 export function onFieldKeyDown(ctx: any) {
 
@@ -129,25 +133,25 @@ export function onFieldKeyDown(ctx: any) {
           break;
         }
         default: {
-          if (isSaveShortcut(event)) {
+          if (isSaveShortcut(event) && geScreenActionButtonsState(ctx)?.isSaveButtonVisible) {
             tablePanelView.setEditing(false);
             yield*flushCurrentRowData(ctx)();
             const formScreenLifecycle = getFormScreenLifecycle(ctx);
             yield*formScreenLifecycle.onSaveSession();
           }
-          else if (isRefreshShortcut(event)) {
+          else if (isRefreshShortcut(event) && geScreenActionButtonsState(ctx)?.isRefreshButtonVisible) {
             tablePanelView.setEditing(false);
             yield*flushCurrentRowData(ctx)();
             const formScreenLifecycle = getFormScreenLifecycle(ctx);
             yield*formScreenLifecycle.onRequestScreenReload();
           }
-          else if (isAddRecordShortcut(event)) {
+          else if (isAddRecordShortcut(event) && getIsAddButtonVisible(dataView)) {
             yield onCreateRowClick(dataView)(event);
           }
-          else if (isDeleteRecordShortcut(event)) {
+          else if (isDeleteRecordShortcut(event) && getIsDelButtonVisible(dataView)) {
             yield onDeleteRowClick(dataView)(event);
           }
-          else if (isDuplicateRecordShortcut(event)) {
+          else if (isDuplicateRecordShortcut(event) && getIsCopyButtonVisible(dataView)) {
             yield onCopyRowClick(dataView)(event);
           }
           else if (isFilterRecordShortcut(event)) {
